p4i integration Integrate 210230 from 4.1 to main: Fixes:    QIODevice::getch(), putch() and ungetch() with non-ascii characters Task:     103139 RevBy:    Simon AutoTest: 210229 Details:  In Qt 3, getch() returned -1 or 0-255. getChar() operates on (signed) 	  chars. QIODevice used a broken cast that resulted in Qt 4's getch(), 	  which is a support function for Qt 3, returning negative values for 	  characters 0x80-0xff.
[git-p4: depot-paths = "//depot/qt/main/": change = 210231]

+++ src/corelib/io/qiodevice.h
@@ -163,9 +163,9 @@ public:
     inline QT3_SUPPORT qint64 writeBlock(const char *data, quint64 len) { return write(data, len); }
     inline QT3_SUPPORT qint64 writeBlock(const QByteArray &data) { return write(data); }
 
-    inline QT3_SUPPORT int getch() { char c; return getChar(&c) ? int(c) : -1; }
-    inline QT3_SUPPORT int putch(int c) { return putChar(c) ? int(c) : -1; }
-    inline QT3_SUPPORT int ungetch(int c) { ungetChar(c); return c; }
+    inline QT3_SUPPORT int getch() { char c; return getChar(&c) ? int(uchar(c)) : -1; }
+    inline QT3_SUPPORT int putch(int c) { return putChar(c) ? int(uchar(c)) : -1; }
+    inline QT3_SUPPORT int ungetch(int c) { ungetChar(uchar(c)); return c; }
 #endif
 };
